Removed calls for changing the fs segment protecting kernel memory
This is most certainly incomplete and needs to be replaced.  This is
just being done for an attempt at compiling the driver and finding other
regressions that are present. With this commit, the driver compiles and
loads but fails to initialize hardware and most certainly needs the code
below fixed.